API reformat

+++ server/src/main/java/com/matzip/server/domain/user/service/UserService.java
@@ -24,7 +24,7 @@ import java.util.Collections;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional(readOnly=true)
 public class UserService {
     private final UserRepository userRepository;
 
@@ -44,56 +44,49 @@ public class UserService {
             throw new UsernameAlreadyExistsException(signUpRequest.getUsername());
         User user = userRepository.save(new User(signUpRequest, passwordEncoder));
         UserPrincipal userPrincipal = new UserPrincipal(user);
-        String token = jwtProvider.generateAccessToken(new MatzipAuthenticationToken(
-                userPrincipal, null, userPrincipal.getAuthorities()));
+        String token = jwtProvider.generateAccessToken(
+                new MatzipAuthenticationToken(userPrincipal, null, userPrincipal.getAuthorities()));
         return new UserDto.SignUpResponse(new UserDto.Response(user), token);
     }
 
     public UserDto.Response findUser(UserDto.FindRequest findRequest, String userRole) {
-        User user = userRepository.findByUsername(findRequest.getUsername()).orElseThrow(
-                () -> new UsernameNotFoundException(findRequest.getUsername()));
+        User user = userRepository.findByUsername(findRequest.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException(findRequest.getUsername()));
         if (user.getRole().equals("ADMIN") && !userRole.equals("ADMIN"))
             throw new AdminUserAccessByNormalUserException();
         return new UserDto.Response(user);
     }
 
     public Page<UserDto.Response> searchUsers(UserDto.SearchRequest searchRequest) {
-        PageRequest pageRequest = PageRequest.of(
-                searchRequest.getPageNumber(),
-                searchRequest.getPageSize(),
-                Sort.by("createdAt").ascending());
-        Page<User> users = userRepository
-                .findAllByUsernameContainsIgnoreCaseAndIsNonLockedTrueAndRoleEquals(pageRequest, searchRequest.getUsername(), "NORMAL");
+        PageRequest pageRequest = PageRequest.of(searchRequest.getPageNumber(), searchRequest.getPageSize(),
+                                                 Sort.by("createdAt").ascending());
+        Page<User> users = userRepository.findAllByUsernameContainsIgnoreCaseAndIsNonLockedTrueAndRoleEquals(
+                pageRequest, searchRequest.getUsername(), "NORMAL");
         return users.map(UserDto.Response::new);
     }
 
     @Transactional
     public void changePassword(UserDto.PasswordChangeRequest passwordChangeRequest) {
-        User user = userRepository.findByUsername(passwordChangeRequest.getUsername()).orElseThrow(
-                () -> new UsernameNotFoundException(passwordChangeRequest.getUsername())
-        );
+        User user = userRepository.findByUsername(passwordChangeRequest.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException(passwordChangeRequest.getUsername()));
         userRepository.save(user.changePassword(passwordChangeRequest, passwordEncoder));
     }
 
     public UserDto.Response getMe(String username) {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException(username));
+        User user = userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException(username));
         return new UserDto.Response(user);
     }
 
     @Transactional
     public void deleteMe(Long id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new UserIdNotFoundException(id));
-        if (user.getRole().equals("ADMIN"))
-            throw new DeleteAdminUserException();
+        User user = userRepository.findById(id).orElseThrow(() -> new UserIdNotFoundException(id));
+        if (user.getRole().equals("ADMIN")) throw new DeleteAdminUserException();
         userRepository.delete(user);
     }
 
     @Transactional
     public UserDto.Response patchMe(String username, UserDto.ModifyProfileRequest modifyProfileRequest) {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException(username));
+        User user = userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException(username));
         if (modifyProfileRequest.getProfileImageUrl() != null)
             imageService.deleteImages(username, Collections.singletonList(user.getProfileImageUrl()));
         userRepository.save(user.patch(modifyProfileRequest));
